Making a lib project with a bin file to run it.
