+++ src/lib.rs
@@ -1,0 +1,20 @@
+extern crate config;
+#[macro_use]
+extern crate diesel;
+#[macro_use]
+extern crate error_chain;
+extern crate mio;
+extern crate serde;
+#[macro_use]
+extern crate serde_derive;
+extern crate sysfs_gpio;
+
+pub mod errors;
+pub mod configuration;
+pub mod database;
+pub mod great_manager;
+pub mod key_mapper;
+pub mod rfid_reader;
+pub mod rfid_buffer;
+pub mod schema;
+pub mod models;
